backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
flask-jwt-extended==4.6.0
openai-whisper==20231117
transformers==4.40.0
sentencepiece==0.2.0
gTTS==2.5.1
moviepy==1.0.3
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
eventlet==0.36.1
werkzeug==3.0.3
sacremoses==0.1.1
psycopg2-binary==2.9.9
bcrypt==4.1.3
